refactor(charts): extract scale helpers into ChartScaleMixin
Move the scale and axis-data construction helpers out of the Chart base
class into a new charted/charts/_chart_scales.py module. The methods
(_reject_unsupported_scales, _is_time_scale, _normalize_time_data,
_anchor_axis_data, _build_grid_config, _build_scale) and the
_BAR_VALUE_AXIS class attribute move verbatim and resolve back onto Chart
via the MRO, so __init__ and rendering are unchanged.

The read-only x_scale/y_scale property accessors stay on Chart. Drop the
now-unused Scale import and a type: ignore[misc] in combo.py that mypy
flags as redundant once the bases resolve.
